requirements.txt
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.1.0
autopep8==1.4.4
certifi==2019.3.9
chardet==3.0.4
Click==7.0
decorator==4.4.0
dnspython==1.16.0
docopt==0.6.2
eventlet==0.24.1
Flask==1.0.2
greenlet==0.4.15
h5py==2.9.0
idna==2.8
imageio==2.5.0
imageio-ffmpeg==0.3.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
monotonic==1.5
moviepy==1.0.0
multidict==4.5.2
numpy==1.16.3
pandas==0.24.2
Pillow==6.0.0
proglog==0.1.9
ptyprocess==0.6.0
pycodestyle==2.5.0
python-dateutil==2.8.0
python-engineio==3.5.1
python-socketio==4.0.1
pytz==2019.1
requests==2.22.0
six==1.12.0
terminado==0.8.2
tornado==5.1.1
tqdm==4.32.1
urllib3==1.25.2
virtualenv==16.5.0
Werkzeug==0.15.4
yarl==1.3.0
